refactor: 혼동 방지를 위한 컴포넌트 파일명 변경 page -> ShoeCloset
