+++ frontend/src/router/index.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter } from "react-router-dom";
 import Layout from "@components/layout/Layout";
 import NotFound from "@pages/NotFound/NotFound";
-import ShoeCloset from "@pages/ShoeCloset/page";
+import ShoeCloset from "@/pages/ShoeCloset/ShoeCloset";
 import Chat from "@pages/Chat/Chat";
 import Review from "@/pages/Review/Review";
 import MyPage from "@pages/MyPage/MyPage";
